Eliminate Col's reverse property
- Col's reverse property was a boolean that would apply the 'reverse' class if true.  However, there is no support for reverse for columns in flexboxgrid.

There is a .col.reverse class defined, but it sets the flex-direction to column-reverse, which has no effect on an element that is not displayed as flex.

Also, columns are not given the .col class in the way that rows are given the .row class, so the designation of .reverse on a Col would not even cause these rules to be applied.

The .col.reverse class appears to be lint, left over from an abandoned attempt to support columnar flexboxes in the flexboxgrid package.  See:  https://github.com/kristoferjoseph/flexboxgrid/issues/173